+++ app/decorators/organization_decorator.rb
@@ -1,38 +1,11 @@
 # frozen_string_literal: true
 
 class OrganizationDecorator < ApplicationDecorator
+  include PersonOrgDecoratorCommon
+
   def db_detail_link(options = {})
     name = options.delete(:name).presence || self.name
     h.link_to name, h.edit_db_organization_path(self), options
-  end
-
-  def local_name
-    return name if I18n.locale == :ja
-    return name_en if name_en.present?
-    name
-  end
-
-  def name_link
-    h.link_to name, h.organization_path(self)
-  end
-
-  def twitter_username_link
-    url = "https://twitter.com/#{twitter_username}"
-    h.link_to "@#{twitter_username}", url, target: "_blank"
-  end
-
-  def wikipedia_url_link
-    h.link_to "Wikipedia", wikipedia_url, target: "_blank"
-  end
-
-  def url_link
-    h.link_to URI.parse(url).host.downcase, url, target: "_blank"
-  end
-
-  def local_name
-    return name if I18n.locale == :ja
-    return name_en if name_en.present?
-    name
   end
 
   def to_values
