+++ app/models/organization.rb
@@ -3,15 +3,19 @@
 #
 # Table name: organizations
 #
-#  id               :integer          not null, primary key
-#  name             :string           not null
-#  url              :string
-#  wikipedia_url    :string
-#  twitter_username :string
-#  aasm_state       :string           default("published"), not null
-#  created_at       :datetime         not null
-#  updated_at       :datetime         not null
-#  name_kana        :string           default(""), not null
+#  id                  :integer          not null, primary key
+#  name                :string           not null
+#  url                 :string
+#  wikipedia_url       :string
+#  twitter_username    :string
+#  aasm_state          :string           default("published"), not null
+#  created_at          :datetime         not null
+#  updated_at          :datetime         not null
+#  name_kana           :string           default(""), not null
+#  name_en             :string           default(""), not null
+#  url_en              :string           default(""), not null
+#  wikipedia_url_en    :string           default(""), not null
+#  twitter_username_en :string           default(""), not null
 #
 # Indexes
 #
