+++ app/views/organizations/_sidebar.html.slim
@@ -1,0 +1,24 @@
+h1.h2 = organization.decorate.local_name
+span.tag.u-tag-organization
+  = t "words.organization"
+h2.h4.font-weight-bold.mt-2.mb-1
+  = icon "sticky-note-o", class: "mr-zp25"
+  = t "words.information"
+dl
+  dt.small = t "words.name_kanji"
+  dd = organization.name
+  - if organization.name_kana.present?
+    dt.small = t "words.name_kana"
+    dd = organization.name_kana
+  - if organization.name_en.present?
+    dt.small = t "words.name_en"
+    dd = organization.name_en
+  - if organization.url.present?
+    dt.small = t "words.official_site"
+    dd = link_with_domain organization.url
+  - if organization.twitter_username.present?
+    dt.small = t "words.twitter"
+    dd = organization.decorate.twitter_username_link
+  - if organization.wikipedia_url.present?
+    dt.small = t "words.wikipedia"
+    dd = link_with_domain organization.wikipedia_url
