+++ app/views/organizations/show.html+mobile.slim
@@ -1,58 +1,41 @@
-- title = @organization.name
-- description = meta_description("#{title}の情報、参加作品をチェック！")
+- title = t "head.title.organizations.show", title: @organization.decorate.name_with_other_name
+- description = meta_description(t("head.meta.description.organizations.show", title: title))
 - keywords = meta_keywords(@organization.name)
-- set_meta_tags(title: title, description: description, keywords: keywords, reverse: true)
+- set_meta_tags(title: title, description: description, keywords: keywords, og: { title: title, description: description }, reverse: true, prefix: false)
+- set_page_category "organizations_show"
 
-= render "application/v1/navbar"
-
-.content.header
+.c-card.py-1.mb-1
   .container
-    h1
-      .name = @organization.name
+    .text-xs-center
+      h1.h2 = @organization.decorate.local_name
+      span.tag.u-tag-organization
+        = t "words.organization"
 
-.content.info
-  .container
-    h2.title 基本情報
-    table
-      tbody
-        tr
-          th 名前
-          td = @organization.name
+    .mt-2
+      h2.h4.font-weight-bold.text-xs-center.mt-2.mb-1
+        = icon "sticky-note-o", class: "mr-zp25"
+        = t "words.information"
+      dl.row.mb-0
+        dt.col-xs-6.text-xs-right = t "words.name_kanji"
+        dd.col-xs-6 = @organization.name
+        - if @organization.name_kana.present?
+          dt.col-xs-6.text-xs-right = t "words.name_kana"
+          dd.col-xs-6 = @organization.name_kana
+        - if @organization.name_en.present?
+          dt.col-xs-6.text-xs-right = t "words.name_en"
+          dd.col-xs-6 = @organization.name_en
         - if @organization.url.present?
-          tr
-            th 公式サイト
-            td = @organization.decorate.url_link
+          dt.col-xs-6.text-xs-right = t "words.official_site"
+          dd.col-xs-6 = link_with_domain @organization.url
         - if @organization.twitter_username.present?
-          tr
-            th Twitter
-            td = @organization.decorate.twitter_username_link
+          dt.col-xs-6.text-xs-right = t "words.twitter"
+          dd.col-xs-6 = @organization.decorate.twitter_username_link
         - if @organization.wikipedia_url.present?
-          tr
-            th Wikipedia
-            td = @organization.decorate.wikipedia_url_link
-    
-    .edit-link
-      - if user_signed_in? && !browser.device.mobile?
-        - if current_user.committer?
-          = link_to "編集する", edit_db_organization_path(@organization)
-        - else
-          = link_to "編集する", new_db_draft_organization_path(organization_id: @organization.id)
+          dt.col-xs-6.text-xs-right = t "words.wikipedia"
+          dd.col-xs-6 = link_with_domain @organization.wikipedia_url
 
-.content.works
-  .container
-    h2.title 参加作品
-    - @staff_years.each do |year|
-      h3.title-year = "#{year}年"
-      ul
-        - @staffs_with_year[year].group_by(&:work_id).each do |_, wos|
-          - work = wos.first.work
-          li
-            .left
-              = link_to work_path(work) do
-                = ann_image_tag(work.item, :tombo_image, msize: "60x60", size: "140x140", alt: work.title)
-            .right
-              = link_to work_path(work) do
-                .work-title = work.title
-              .wo-role = wos.map { |wo| wo.decorate.role_name }.join("、")
+.container.py-1
+  = render "people/grid", resource_years: @staff_years, resources_with_year: @staffs_with_year
 
-= render "application/v2/components/footer"
+= render "application/components/footer"
+= render "application/components/tabbar"
