+++ app/views/people/_grid.html.slim
@@ -14,4 +14,4 @@
             .text-truncate
               = link_to work.decorate.local_title, work_path(work)
             .text-truncate.small
-              = resources.map { |resource| resource.decorate.part }.join(", ")
+              = resources.map { |resource| resource.instance_of?(Cast) ? resource.decorate.part : resource.decorate.role_name }.join(", ")
